# backend/internal/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

import (
	"time"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// CreatePromptRequest defines model for CreatePromptRequest.
type CreatePromptRequest = PromptCreateRequest

// CreateTranscriptionRequest defines model for CreateTranscriptionRequest.
type CreateTranscriptionRequest = TranscriptionCreateRequest

// HealthResponse defines model for HealthResponse.
type HealthResponse = Health

// PromptResponse defines model for PromptResponse.
type PromptResponse = Prompt

// PromptsResponse defines model for PromptsResponse.
type PromptsResponse = PromptListResponse

// TranscriptionResponse defines model for TranscriptionResponse.
type TranscriptionResponse = Transcription

// TranscriptionsResponse defines model for TranscriptionsResponse.
type TranscriptionsResponse = TranscriptionListResponse

// UpdatePromptRequest defines model for UpdatePromptRequest.
type UpdatePromptRequest = PromptUpdateRequest

// UserResponse defines model for UserResponse.
type UserResponse = User

// Health defines model for health.
type Health struct {
	Status string `json:"status"`
}

// Prompt defines model for prompt.
type Prompt struct {
	Body      string    `json:"body"`
	CreatedAt time.Time `json:"created_at"`
	Id        int64     `json:"id"`
	IsActive  bool      `json:"is_active"`
	IsDefault bool      `json:"is_default"`
	Name      string    `json:"name"`
	UpdatedAt time.Time `json:"updated_at"`
	UserId    *int64    `json:"user_id,omitempty"`
}

// PromptCreateRequest defines model for prompt_create_request.
type PromptCreateRequest struct {
	Body     string `json:"body"`
	IsActive *bool  `json:"is_active,omitempty"`
	Name     string `json:"name"`
}

// PromptListResponse defines model for prompt_list_response.
type PromptListResponse = []Prompt

// PromptUpdateRequest defines model for prompt_update_request.
type PromptUpdateRequest struct {
	Body     *string `json:"body,omitempty"`
	IsActive *bool   `json:"is_active,omitempty"`
	Name     *string `json:"name,omitempty"`
}

// Transcription defines model for transcription.
type Transcription struct {
	CreatedAt    time.Time                `json:"created_at"`
	FullText     string                   `json:"full_text"`
	Id           int64                    `json:"id"`
	SegmentsJson []map[string]interface{} `json:"segments_json"`
	UpdatedAt    time.Time                `json:"updated_at"`
	UserId       int64                    `json:"user_id"`
}

// TranscriptionCreateRequest defines model for transcription_create_request.
type TranscriptionCreateRequest struct {
	FullText     string                   `json:"full_text"`
	SegmentsJson []map[string]interface{} `json:"segments_json"`
}

// TranscriptionListResponse defines model for transcription_list_response.
type TranscriptionListResponse = []Transcription

// User defines model for user.
type User struct {
	AuthProvider string  `json:"auth_provider"`
	Email        *string `json:"email,omitempty"`
	Id           int64   `json:"id"`
	Name         *string `json:"name,omitempty"`
}

// CreatePromptJSONRequestBody defines body for CreatePrompt for application/json ContentType.
type CreatePromptJSONRequestBody = PromptCreateRequest

// UpdatePromptJSONRequestBody defines body for UpdatePrompt for application/json ContentType.
type UpdatePromptJSONRequestBody = PromptUpdateRequest

// CreateTranscriptionJSONRequestBody defines body for CreateTranscription for application/json ContentType.
type CreateTranscriptionJSONRequestBody = TranscriptionCreateRequest